scripts/requirements.txt
overrides
pydantic<2.0
fastapi<0.100.0,>=0.95.2
typing-extensions
openai
tiktoken
numpy<2.0
tenacity
python-dotenv
weaviate-client
pdfplumber
python-docx
tqdm